main_algorithm_v2/offline/requirements.txt
torch
numpy
scipy
scikit-learn
tqdm
pydantic
PyYAML 